restructuring to comply to ROS standards
