+++ iarc_roombafilter/scripts/f_ros.py
@@ -22,11 +22,16 @@ from iarc_main.msg import Roomba, RoombaList, StartRound
 from sensor_msgs.msg import CameraInfo
 from std_msgs.msg import Header, Bool
 
-import f_config as cfg
-from f_manager import UKFManager
-# Filter-Related ...
-from f_utils import *
-from f_model import TargetRoombaModel, ObstacleRoombaModel
+from iarc_roombafilter import f_config as cfg
+from iarc_roombafilter.f_manager import UKFManager
+from iarc_roombafilter.f_utils import *
+from iarc_roombafilter.f_model import TargetRoombaModel, ObstacleRoombaModel
+
+#import f_config as cfg
+#from f_manager import UKFManager
+## Filter-Related ...
+#from f_utils import *
+#from f_model import TargetRoombaModel, ObstacleRoombaModel
 
 
 class UKFManagerROS(object):
